// src/store/requestStore/action.enum.ts
enum actionTypes {
  GET_REQUEST_PENDING = "GET_REQUEST_PENDING",
  GET_REQUEST_SUCCESS = "GET_REQUEST_SUCCESS",
  GET_REQUEST_FAILED = "GET_REQUEST_FAILED",

  NEW_REQUEST_PENDING = "NEW_REQUEST_PENDING",
  NEW_REQUEST_SUCCESS = "NEW_REQUEST_SUCCESS",
  NEW_REQUEST_FAILED = "NEW_REQUEST_FAILED",

  GET_SINGLE_REQUEST_PENDING = "GET_SINGLE_REQUEST_PENDING",
  GET_SINGLE_REQUEST_SUCCESS = "GET_SINGLE_REQUEST_SUCCESS",
  GET_SINGLE_REQUEST_FAILED = "GET_SINGLE_REQUEST_FAILED",

  UPDATE_REQUEST_PENDING = "UPDATE_REQUEST_PENDING",
  UPDATE_REQUEST_SUCCESS = "UPDATE_REQUEST_SUCCESS",
  UPDATE_REQUEST_FAILED = "UPDATE_REQUEST_FAILED",

  DELETE_REQUEST_PENDING = "DELETE_REQUEST_PENDING",
  DELETE_REQUEST_SUCCESS = "DELETE_REQUEST_SUCCESS",
  DELETE_REQUEST_FAILED = "DELETE_REQUEST_FAILED",
}
export default actionTypes;